moved paper validation to paper package
