+++ handlers/new-product-api/src/test/scala/com/gu/newproduct/api/addsubscription/validation/voucher/PaperAccountValidationTest.scala
@@ -1,6 +1,7 @@
 package com.gu.newproduct.api.addsubscription.validation.voucher
 
 import com.gu.i18n.Currency.{GBP, USD}
+import com.gu.newproduct.api.addsubscription.validation.paper.PaperAccountValidation
 import com.gu.newproduct.api.addsubscription.validation.{Failed, Passed, ValidatedAccount}
 import com.gu.newproduct.api.addsubscription.zuora.GetAccount.{AccountBalanceMinorUnits, AutoPay, PaymentMethodId, SfContactId}
 import org.scalatest.{FlatSpec, Matchers}
